Time: 7 ms (13.96%), Space: 44 MB (41.21%) - LeetHub

+++ 0201-bitwise-and-of-numbers-range/0201-bitwise-and-of-numbers-range.java
@@ -8,7 +8,7 @@ class Solution {
         }
         int mask = 0;
         for(int i = 30;i>=diffBits;i--){
-            mask+=(long)Math.pow(2,i);
+            mask+=Math.pow(2,i);
         }
         return left & right & mask;
     }
